refactor form controls inputs to be in separate files
