+++ client/src/components/form/form.css
@@ -67,7 +67,7 @@ button[type="submit"]  {
     border-radius: 20px;
     border: 1px solid #999;
     background: #a4bfc7;
-    padding: 10px 10px;
+    padding: 8px 10px;
     min-width: 100px;
     letter-spacing: 2px;
     text-transform: uppercase;
@@ -75,7 +75,7 @@ button[type="submit"]  {
 
 input {
     border-radius: 20px;
-    border: 1px solid #999;
+    border: 1px solid #ccc;
     background: #fff;
     padding: 8px 10px;
     display: inline-block;
@@ -89,6 +89,6 @@ input {
 .select-input {
     padding: 5px;
     border-radius: 20px;
-    border: 1px solid #999;
+    border: 1px solid #ccc;
     display: inline-block;
 }
