--- client/src/custom-form/components/form-sample.js
@@ -1,15 +1,0 @@
-import React from 'react'
-
-import {
-    Form,
-    ControlInput,
-    ControlSelect,
-    ControlRadioOrCheckbox } from '../../components/form'
-
-import '../../components/form/form.css';
-
-export default (props) => (
-
-
-
-)
